requirements.txt
youtube-transcript-api
langchain
langchain_core
python-dotenv
python-docx
sentencepiece
torch
transformers
nltk==3.8.1
openai==0.28.0
gTTS
requests
clip
textwrap3
Pillow
moviepy
flask
beautifulsoup4==4.12.2
matplotlib==3.7.1
moviepy==1.0.3
numpy
selenium==4.12.0
yt_dlp==2023.9.24